honor/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.1.2
django-ckeditor==5.6.1
django-js-asset==1.1.0
idna==2.7
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
